Update header file and function

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -8,11 +8,11 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	while (h)
 	{
-		printf("%ld\n", h->n);
+		printf("%d\n", h->n);
 		n++;
 		h = h->next;
 	}
